rename option_results.py to extracted_feature.py
